@@ -1,22 +1,20 @@
 import 'package:flutter/material.dart';
 
-class Resposta  extends StatelessWidget {
-  final String textButton;
-  final void Function() quandoSelecionado;
+class Answer extends StatelessWidget {
+  final String text;
+  final void Function() onSelect;
 
-  Resposta(this.textButton, this.quandoSelecionado);
-
+  Answer(this.text, this.onSelect);
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: double.infinity,
-      child: ElevatedButton(
-        // color = Colors.blue,
-        // textColor = Colors.white,
-        child: Text(textButton),
-        onPressed: quandoSelecionado,
-      ),
-    );
+        width: 200,
+        child: ElevatedButton(
+            style: ElevatedButton.styleFrom(primary: Colors.blue),
+            onPressed: onSelect,
+            child: Text(
+              text,
+              style: TextStyle(color: Colors.white),
+            )));
   }
-
 }
